Create Cash Deposit Prediction
The data given is of cash deposits made in particular branches of a bank. Please go through the descriptions of variables as below for a conceptional understanding of the data.

The description of the features is given below :

Feature	Description
Serial Number	Unique identifier for each data point relationship
Main Office	Is the branch where the deposit was made the main office of the bank. 0 signifies no and 1 signifies yes
Branch Number	Internally used identification number of the branch or certain branches
Establishment Date and Acquired Date	The date at which the bank branch was established and acquired by the specific bank franchise
City, Country and State	Geographical Details of the bank branch
Year Deposits Variables( 2013 Deposits, 2014 Deposits…)	Deposit amount in U.S. dollars for specific branch for a specific year